builder restructure.
