+++ app/src/main/java/com/bigkoo/pickerviewdemo/MainActivity.java
@@ -5,7 +5,6 @@ import android.graphics.Color;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
-import android.view.Gravity;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
@@ -16,9 +15,13 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.bigkoo.pickerview.OptionsPickerView;
-import com.bigkoo.pickerview.TimePickerView;
+import com.bigkoo.pickerview.builder.OptionsPickerBuilder;
+import com.bigkoo.pickerview.builder.TimePickerBuilder;
+import com.bigkoo.pickerview.listener.OnOptionsSelectListener;
+import com.bigkoo.pickerview.view.OptionsPickerView;
+import com.bigkoo.pickerview.view.TimePickerView;
 import com.bigkoo.pickerview.listener.CustomListener;
+import com.bigkoo.pickerview.listener.OnTimeSelectListener;
 import com.bigkoo.pickerviewdemo.bean.CardBean;
 import com.bigkoo.pickerviewdemo.bean.ProvinceBean;
 
@@ -33,8 +36,9 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     private ArrayList<ProvinceBean> options1Items = new ArrayList<>();
     private ArrayList<ArrayList<String>> options2Items = new ArrayList<>();
 
-    /*   private ArrayList<ArrayList<ArrayList<IPickerViewData>>> options3Items = new ArrayList<>();*/
-    private Button btn_Time, btn_Options, btn_CustomOptions, btn_CustomTime, btn_no_linkage, btn_to_Fragment;
+    private Button btn_Options;
+    private Button btn_CustomOptions;
+    private Button btn_CustomTime;
 
     private TimePickerView pvTime, pvCustomTime, pvCustomLunar;
     private OptionsPickerView pvOptions, pvCustomOptions, pvNoLinkOptions;
@@ -59,12 +63,12 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         initCustomOptionPicker();
         initNoLinkOptionsPicker();
 
-        btn_Time = (Button) findViewById(R.id.btn_Time);
+        Button btn_Time = (Button) findViewById(R.id.btn_Time);
         btn_Options = (Button) findViewById(R.id.btn_Options);
         btn_CustomOptions = (Button) findViewById(R.id.btn_CustomOptions);
         btn_CustomTime = (Button) findViewById(R.id.btn_CustomTime);
-        btn_no_linkage = (Button) findViewById(R.id.btn_no_linkage);
-        btn_to_Fragment = (Button) findViewById(R.id.btn_fragment);
+        Button btn_no_linkage = (Button) findViewById(R.id.btn_no_linkage);
+        Button btn_to_Fragment = (Button) findViewById(R.id.btn_fragment);
 
         btn_Time.setOnClickListener(this);
         btn_Options.setOnClickListener(this);
@@ -112,7 +116,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         Calendar endDate = Calendar.getInstance();
         endDate.set(2069, 2, 28);
         //时间选择器 ，自定义布局
-        pvCustomLunar = new TimePickerView.Builder(this, new TimePickerView.OnTimeSelectListener() {
+        pvCustomLunar = new TimePickerBuilder(this, new OnTimeSelectListener() {
             @Override
             public void onTimeSelect(Date date, View v) {//选中事件回调
                 Toast.makeText(MainActivity.this, getTime(date), Toast.LENGTH_SHORT).show();
@@ -179,7 +183,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     }
 
 
-    private void initTimePicker() {
+    /*private void initTimePicker() {
         //控制时间范围(如果不设置范围，则使用默认时间1900-2100年，此段代码可注释)
         //因为系统Calendar的月份是从0-11的,所以如果是调用Calendar的set方法来设置时间,月份的范围也要是从0-11
         Calendar selectedDate = Calendar.getInstance();
@@ -188,11 +192,11 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         Calendar endDate = Calendar.getInstance();
         endDate.set(2019, 11, 28);
         //时间选择器
-        pvTime = new TimePickerView.Builder(this, new TimePickerView.OnTimeSelectListener() {
+        pvTime = new TimePickerView.Builder(this, new OnTimeSelectListener() {
             @Override
             public void onTimeSelect(Date date, View v) {//选中事件回调
                 // 这里回调过来的v,就是show()方法里面所添加的 View 参数，如果show的时候没有添加参数，v则为null
-                /*btn_Time.setText(getTime(date));*/
+                btn_Time.setText(getTime(date));
                 Button btn = (Button) v;
                 btn.setText(getTime(date));
             }
@@ -208,7 +212,19 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 //                .setBackgroundId(0x00FFFFFF) //设置外部遮罩颜色
 //                .setDecorView(null)
                 .build();
+    }*/
+
+    private void initTimePicker() {
+
+        pvTime = new TimePickerBuilder(this, new OnTimeSelectListener() {
+            @Override
+            public void onTimeSelect(Date date, View v) {
+
+            }
+        }).build();
+
     }
+
 
     private void initCustomTimePicker() {
 
@@ -227,7 +243,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         Calendar endDate = Calendar.getInstance();
         endDate.set(2027, 2, 28);
         //时间选择器 ，自定义布局
-        pvCustomTime = new TimePickerView.Builder(this, new TimePickerView.OnTimeSelectListener() {
+        pvCustomTime = new TimePickerBuilder(this, new OnTimeSelectListener() {
             @Override
             public void onTimeSelect(Date date, View v) {//选中事件回调
                 btn_CustomTime.setText(getTime(date));
@@ -289,7 +305,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
          * 注意 ：如果是三级联动的数据(省市区等)，请参照 JsonDataActivity 类里面的写法。
          */
 
-        pvOptions = new OptionsPickerView.Builder(this, new OptionsPickerView.OnOptionsSelectListener() {
+        pvOptions = new OptionsPickerBuilder(this, new OnOptionsSelectListener() {
             @Override
             public void onOptionsSelect(int options1, int options2, int options3, View v) {
                 //返回的分别是三个级别的选中位置
@@ -329,7 +345,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
          * 自定义布局中，id为 optionspicker 或者 timepicker 的布局以及其子控件必须要有，否则会报空指针。
          * 具体可参考demo 里面的两个自定义layout布局。
          */
-        pvCustomOptions = new OptionsPickerView.Builder(this, new OptionsPickerView.OnOptionsSelectListener() {
+        pvCustomOptions = new OptionsPickerBuilder(this, new OnOptionsSelectListener() {
             @Override
             public void onOptionsSelect(int options1, int option2, int options3, View v) {
                 //返回的分别是三个级别的选中位置
@@ -377,7 +393,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     }
 
     private void initNoLinkOptionsPicker() {// 不联动的多级选项
-        pvNoLinkOptions = new OptionsPickerView.Builder(this, new OptionsPickerView.OnOptionsSelectListener() {
+        pvNoLinkOptions = new OptionsPickerBuilder(this, new OnOptionsSelectListener() {
 
             @Override
             public void onOptionsSelect(int options1, int options2, int options3, View v) {
