+++ pickerview/src/main/java/com/bigkoo/pickerview/configure/PickerOptions.java
@@ -5,57 +5,51 @@ import android.graphics.Typeface;
 import android.view.Gravity;
 import android.view.ViewGroup;
 
-import com.bigkoo.pickerview.OptionsPickerView;
+import com.bigkoo.pickerview.listener.OnOptionsSelectListener;
 import com.bigkoo.pickerview.R;
-import com.bigkoo.pickerview.TimePickerView;
 import com.bigkoo.pickerview.listener.CustomListener;
+import com.bigkoo.pickerview.listener.OnTimeSelectListener;
 import com.contrarywind.view.WheelView;
 
 import java.util.Calendar;
 
 /**
  * 配置类
- * Created by XiaoSong on 2018/3/8.
+ * Created by xiaosongzeem on 2018/3/8.
  */
 
 public class PickerOptions {
 
     //常量
-    private static final int PICKERVIEW_BTN_NORMAL = 0xFF057dff;
-    private static final int PICKERVIEW_BG_TOPBAR = 0xFFf5f5f5;
-    private static final int PICKERVIEW_TOPBAR_TITLE = 0xFF000000;
-    private static final int PICKERVIEW_BG_COLOR_DEFAULT = 0xFFFFFFFF;
+    private static final int PICKER_VIEW_BTN_COLOR_NORMAL = 0xFF057dff;
+    private static final int PICKER_VIEW_BG_COLOR_TITLE = 0xFFf5f5f5;
+    private static final int PICKER_VIEW_COLOR_TITLE = 0xFF000000;
+    private static final int PICKER_VIEW_BG_COLOR_DEFAULT = 0xFFFFFFFF;
 
     public static final int TYPE_PICKER_OPTIONS = 1;
     public static final int TYPE_PICKER_TIME = 2;
 
 
-    public OptionsPickerView.OnOptionsSelectListener optionsSelectListener;
-    public TimePickerView.OnTimeSelectListener timeSelectListener;
+    public OnOptionsSelectListener optionsSelectListener;
+    public OnTimeSelectListener timeSelectListener;
     public CustomListener customListener;
 
 
     //options picker
-    public String label1;
-    public String label2;
-    public String label3;
+    public String label1, label2, label3;//单位字符
+    public int option1, option2, option3;//默认选中项
+    public int x_offset_one, x_offset_two, x_offset_three;//x轴偏移量
 
     public boolean cyclic1 = false;//是否循环，默认否
     public boolean cyclic2 = false;
     public boolean cyclic3 = false;
 
-    public int option1;//默认选中项
-    public int option2;
-    public int option3;
-
-    public int x_offset_one;//x轴偏移量
-    public int x_offset_two;
-    public int x_offset_three;
     public boolean linkage = true;
     public boolean isRestoreItem = false; //切换时，还原第一项
 
+
     //time picker
-    public boolean[] type = new boolean[]{true, true, true, true, true, true};//显示类型 默认全部显示
+    public boolean[] type = new boolean[]{true, true, true, false, false, false};//显示类型，默认显示： 年月日
 
     public Calendar date;//当前选中时间
     public Calendar startDate;//开始时间
@@ -66,7 +60,7 @@ public class PickerOptions {
     public boolean cyclic = false;//是否循环
     public boolean isLunarCalendar = false;//是否显示农历
 
-    public String label_year, label_month, label_day, label_hours, label_mins, label_seconds;//单位
+    public String label_year, label_month, label_day, label_hours, label_minutes, label_seconds;//单位
     public int x_offset_year, x_offset_month, x_offset_day, x_offset_hours, x_offset_minutes, x_offset_seconds;//单位
 
 
@@ -88,12 +82,12 @@ public class PickerOptions {
     public String textContentCancel;//取消按钮文字
     public String textContentTitle;//标题文字
 
-    public int textColorConfirm = PICKERVIEW_BTN_NORMAL;//确定按钮颜色
-    public int textColorCancel = PICKERVIEW_BTN_NORMAL;//取消按钮颜色
-    public int textColorTitle = PICKERVIEW_TOPBAR_TITLE;//标题颜色
+    public int textColorConfirm = PICKER_VIEW_BTN_COLOR_NORMAL;//确定按钮颜色
+    public int textColorCancel = PICKER_VIEW_BTN_COLOR_NORMAL;//取消按钮颜色
+    public int textColorTitle = PICKER_VIEW_COLOR_TITLE;//标题颜色
 
-    public int bgColorWheel = PICKERVIEW_BG_COLOR_DEFAULT;//滚轮背景颜色
-    public int bgColorTitle = PICKERVIEW_BG_TOPBAR;//标题背景颜色
+    public int bgColorWheel = PICKER_VIEW_BG_COLOR_DEFAULT;//滚轮背景颜色
+    public int bgColorTitle = PICKER_VIEW_BG_COLOR_TITLE;//标题背景颜色
 
     public int textSizeSubmitCancel = 17;//确定取消按钮大小
     public int textSizeTitle = 18;//标题文字大小
@@ -108,7 +102,7 @@ public class PickerOptions {
     public boolean isDialog;//是否是对话框模式
 
     public boolean cancelable = true;//是否能取消
-    public boolean isCenterLabel = true;//是否只显示中间的label
+    public boolean isCenterLabel = false;//是否只显示中间的label,默认每个item都显示
     public Typeface font = Typeface.MONOSPACE;//字体样式
     public WheelView.DividerType dividerType = WheelView.DividerType.FILL;//分隔线类型
