+++ pickerview/src/main/java/com/bigkoo/pickerview/view/TimePickerView.java
@@ -1,20 +1,16 @@
-package com.bigkoo.pickerview;
+package com.bigkoo.pickerview.view;
 
 import android.content.Context;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
+import com.bigkoo.pickerview.R;
 import com.bigkoo.pickerview.configure.PickerOptions;
-import com.bigkoo.pickerview.listener.CustomListener;
-import com.bigkoo.pickerview.view.BasePickerView;
-import com.bigkoo.pickerview.view.WheelTime;
-import com.contrarywind.view.WheelView;
 
 import java.text.ParseException;
 import java.util.Calendar;
@@ -30,274 +26,15 @@ public class TimePickerView extends BasePickerView implements View.OnClickListen
 
     private WheelTime wheelTime; //自定义控件
 
-
     private static final String TAG_SUBMIT = "submit";
     private static final String TAG_CANCEL = "cancel";
 
-    //构造方法
+
     public TimePickerView(PickerOptions pickerOptions) {
         super(pickerOptions.context);
         mPickerOptions = pickerOptions;
         initView(pickerOptions.context);
     }
-
-
-    //建造器
-    public static class Builder {
-
-        //配置类
-        private PickerOptions mPickerOptions;
-
-        //Required
-        public Builder(Context context, OnTimeSelectListener listener) {
-            mPickerOptions = new PickerOptions(PickerOptions.TYPE_PICKER_TIME);
-            mPickerOptions.context = context;
-            mPickerOptions.timeSelectListener = listener;
-        }
-
-        //Option
-        public Builder setType(boolean[] type) {
-            mPickerOptions.type = type;
-            return this;
-        }
-
-        public Builder setGravity(int gravity) {
-            mPickerOptions.textGravity = gravity;
-            return this;
-        }
-
-        public Builder setSubmitText(String textContentConfirm) {
-            mPickerOptions.textContentConfirm = textContentConfirm;
-            return this;
-        }
-
-        public Builder isDialog(boolean isDialog) {
-            mPickerOptions.isDialog = isDialog;
-            return this;
-        }
-
-        public Builder setCancelText(String textContentCancel) {
-            mPickerOptions.textContentCancel = textContentCancel;
-            return this;
-        }
-
-        public Builder setTitleText(String textContentTitle) {
-            mPickerOptions.textContentTitle = textContentTitle;
-            return this;
-        }
-
-        public Builder setSubmitColor(int textColorConfirm) {
-            mPickerOptions.textColorConfirm = textColorConfirm;
-            return this;
-        }
-
-        public Builder setCancelColor(int textColorCancel) {
-            mPickerOptions.textColorCancel = textColorCancel;
-            return this;
-        }
-
-        /**
-         * ViewGroup 类型的容器
-         *
-         * @param decorView 选择器会被添加到此容器中
-         * @return this
-         */
-        public Builder setDecorView(ViewGroup decorView) {
-            mPickerOptions.decorView = decorView;
-            return this;
-        }
-
-        public Builder setBgColor(int bgColorWheel) {
-            mPickerOptions.bgColorWheel = bgColorWheel;
-            return this;
-        }
-
-        public Builder setTitleBgColor(int bgColorTitle) {
-            mPickerOptions.bgColorTitle = bgColorTitle;
-            return this;
-        }
-
-        public Builder setTitleColor(int textColorTitle) {
-            mPickerOptions.textColorTitle = textColorTitle;
-            return this;
-        }
-
-        public Builder setSubCalSize(int textSizeSubmitCancel) {
-            mPickerOptions.textSizeSubmitCancel = textSizeSubmitCancel;
-            return this;
-        }
-
-        public Builder setTitleSize(int textSizeTitle) {
-            mPickerOptions.textSizeTitle = textSizeTitle;
-            return this;
-        }
-
-        public Builder setContentTextSize(int textSizeContent) {
-            mPickerOptions.textSizeContent = textSizeContent;
-            return this;
-        }
-
-        /**
-         * 因为系统Calendar的月份是从0-11的,所以如果是调用Calendar的set方法来设置时间,月份的范围也要是从0-11
-         *
-         * @param date
-         * @return
-         */
-        public Builder setDate(Calendar date) {
-            mPickerOptions.date = date;
-            return this;
-        }
-
-        public Builder setLayoutRes(int res, CustomListener customListener) {
-            mPickerOptions.layoutRes = res;
-            mPickerOptions.customListener = customListener;
-            return this;
-        }
-
-        /**
-         * use the setRangDate method instead.
-         *
-         * @deprecated Use {@link  #setRangDate()} with two Calendar value.
-         */
-        @Deprecated
-        public Builder setRange(int startYear, int endYear) {
-            mPickerOptions.startYear = startYear;
-            mPickerOptions.endYear = endYear;
-            return this;
-        }
-
-        /**
-         * 设置起始时间
-         * 因为系统Calendar的月份是从0-11的,所以如果是调用Calendar的set方法来设置时间,月份的范围也要是从0-11
-         */
-
-        public Builder setRangDate(Calendar startDate, Calendar endDate) {
-            mPickerOptions.startDate = startDate;
-            mPickerOptions.endDate = endDate;
-            return this;
-        }
-
-
-        /**
-         * 设置间距倍数,但是只能在1.0-4.0f之间
-         *
-         * @param lineSpacingMultiplier
-         */
-        public Builder setLineSpacingMultiplier(float lineSpacingMultiplier) {
-            mPickerOptions.lineSpacingMultiplier = lineSpacingMultiplier;
-            return this;
-        }
-
-        /**
-         * 设置分割线的颜色
-         *
-         * @param dividerColor
-         */
-        public Builder setDividerColor(int dividerColor) {
-            mPickerOptions.dividerColor = dividerColor;
-            return this;
-        }
-
-        /**
-         * 设置分割线的类型
-         *
-         * @param dividerType
-         */
-        public Builder setDividerType(WheelView.DividerType dividerType) {
-            mPickerOptions.dividerType = dividerType;
-            return this;
-        }
-
-        /**
-         * //显示时的外部背景色颜色,默认是灰色
-         *
-         * @param backgroundId
-         */
-
-        public Builder setBackgroundId(int backgroundId) {
-            mPickerOptions.backgroundId = backgroundId;
-            return this;
-        }
-
-        /**
-         * 设置分割线之间的文字的颜色
-         *
-         * @param textColorCenter
-         */
-        public Builder setTextColorCenter(int textColorCenter) {
-            mPickerOptions.textColorCenter = textColorCenter;
-            return this;
-        }
-
-        /**
-         * 设置分割线以外文字的颜色
-         *
-         * @param textColorOut
-         */
-        public Builder setTextColorOut(int textColorOut) {
-            mPickerOptions.textColorOut = textColorOut;
-            return this;
-        }
-
-        public Builder isCyclic(boolean cyclic) {
-            mPickerOptions.cyclic = cyclic;
-            return this;
-        }
-
-        public Builder setOutSideCancelable(boolean cancelable) {
-            mPickerOptions.cancelable = cancelable;
-            return this;
-        }
-
-        public Builder setLunarCalendar(boolean lunarCalendar) {
-            mPickerOptions.isLunarCalendar = lunarCalendar;
-            return this;
-        }
-
-
-        public Builder setLabel(String label_year, String label_month, String label_day, String label_hours, String label_mins, String label_seconds) {
-            mPickerOptions.label_year = label_year;
-            mPickerOptions.label_month = label_month;
-            mPickerOptions.label_day = label_day;
-            mPickerOptions.label_hours = label_hours;
-            mPickerOptions.label_mins = label_mins;
-            mPickerOptions.label_seconds = label_seconds;
-            return this;
-        }
-
-        /**
-         * 设置X轴倾斜角度[ -90 , 90°]
-         *
-         * @param x_offset_year    年
-         * @param x_offset_month   月
-         * @param x_offset_day     日
-         * @param x_offset_hours   时
-         * @param x_offset_minutes 分
-         * @param x_offset_seconds 秒
-         * @return
-         */
-        public Builder setTextXOffset(int x_offset_year, int x_offset_month, int x_offset_day,
-                                      int x_offset_hours, int x_offset_minutes, int x_offset_seconds) {
-            mPickerOptions.x_offset_year = x_offset_year;
-            mPickerOptions.x_offset_month = x_offset_month;
-            mPickerOptions.x_offset_day = x_offset_day;
-            mPickerOptions.x_offset_hours = x_offset_hours;
-            mPickerOptions.x_offset_minutes = x_offset_minutes;
-            mPickerOptions.x_offset_seconds = x_offset_seconds;
-            return this;
-        }
-
-        public Builder isCenterLabel(boolean isCenterLabel) {
-            mPickerOptions.isCenterLabel = isCenterLabel;
-            return this;
-        }
-
-
-        public TimePickerView build() {
-            return new TimePickerView(mPickerOptions);
-        }
-    }
-
 
     private void initView(Context context) {
         setDialogOutSideCancelable();
@@ -366,7 +103,7 @@ public class TimePickerView extends BasePickerView implements View.OnClickListen
 
         setTime();
         wheelTime.setLabels(mPickerOptions.label_year, mPickerOptions.label_month, mPickerOptions.label_day
-                , mPickerOptions.label_hours, mPickerOptions.label_mins, mPickerOptions.label_seconds);
+                , mPickerOptions.label_hours, mPickerOptions.label_minutes, mPickerOptions.label_seconds);
         wheelTime.setTextXOffset(mPickerOptions.x_offset_year, mPickerOptions.x_offset_month, mPickerOptions.x_offset_day,
                 mPickerOptions.x_offset_hours, mPickerOptions.x_offset_minutes, mPickerOptions.x_offset_seconds);
 
@@ -469,7 +206,8 @@ public class TimePickerView extends BasePickerView implements View.OnClickListen
 
 
     /**
-     *  目前暂时只支持设置1900 - 2100年
+     * 目前暂时只支持设置1900 - 2100年
+     *
      * @param lunar 农历的开关
      */
     public void setLunarCalendar(boolean lunar) {
@@ -486,7 +224,7 @@ public class TimePickerView extends BasePickerView implements View.OnClickListen
 
             wheelTime.setLunarCalendar(lunar);
             wheelTime.setLabels(mPickerOptions.label_year, mPickerOptions.label_month, mPickerOptions.label_day,
-                    mPickerOptions.label_hours, mPickerOptions.label_mins, mPickerOptions.label_seconds);
+                    mPickerOptions.label_hours, mPickerOptions.label_minutes, mPickerOptions.label_seconds);
             wheelTime.setPicker(year, month, day, hours, minute, seconds);
         } catch (ParseException e) {
             e.printStackTrace();
@@ -497,10 +235,6 @@ public class TimePickerView extends BasePickerView implements View.OnClickListen
         return wheelTime.isLunarCalendar();
     }
 
-
-    public interface OnTimeSelectListener {
-        void onTimeSelect(Date date, View v);
-    }
 
     @Override
     public boolean isDialog() {
